Tried to cleanup the models situation. This was largely achieved in three separate steps.
First, all the private types were removed. Using a dataPoint to communicate
but a DataPoint in the API didn't seem particularly useful.

Arrays of pointers when you're already dealing with a slice is just unecessary
indirection.

Finally, tried to clean up the variable names. TS -> Timestamp, TZ -> Timezone
and so on.